+++ common/dialer/dialer.go
@@ -62,7 +62,7 @@ func NewWithOptions(options Options) (N.Dialer, error) {
 			return nil, err
 		}
 	}
-	if options.RemoteIsDomain && (!hasDetour || options.ResolverOnDetour || dialOptions.DomainResolver != nil && dialOptions.DomainResolver.Server != "") {
+	if options.RemoteIsDomain && (!hasDetour || options.ResolverOnDetour || dialOptions.DomainResolver != nil && len(dialOptions.DomainResolver.Server) > 0) {
 		networkManager := service.FromContext[adapter.NetworkManager](options.Context)
 		dnsTransport := service.FromContext[adapter.DNSTransportManager](options.Context)
 		var defaultOptions adapter.NetworkOptions
@@ -70,17 +70,20 @@ func NewWithOptions(options Options) (N.Dialer, error) {
 			defaultOptions = networkManager.DefaultOptions()
 		}
 		var (
-			server               string
+			servers              []string
 			dnsQueryOptions      adapter.DNSQueryOptions
 			resolveFallbackDelay time.Duration
 		)
-		if dialOptions.DomainResolver != nil && dialOptions.DomainResolver.Server != "" {
-			var transport adapter.DNSTransport
+		if dialOptions.DomainResolver != nil && len(dialOptions.DomainResolver.Server) > 0 {
+			var transports []adapter.DNSTransport
 			if !options.DirectResolver {
-				var loaded bool
-				transport, loaded = dnsTransport.Transport(dialOptions.DomainResolver.Server)
-				if !loaded {
-					return nil, E.New("domain resolver not found: " + dialOptions.DomainResolver.Server)
+				transports = make([]adapter.DNSTransport, 0, len(dialOptions.DomainResolver.Server))
+				for _, server := range dialOptions.DomainResolver.Server {
+					transport, loaded := dnsTransport.Transport(server)
+					if !loaded {
+						return nil, E.New("domain resolver not found: ", server)
+					}
+					transports = append(transports, transport)
 				}
 			}
 			var strategy C.DomainStrategy
@@ -93,9 +96,9 @@ func NewWithOptions(options Options) (N.Dialer, error) {
 				strategy = C.DomainStrategy(dialOptions.DomainStrategy)
 				deprecated.Report(options.Context, deprecated.OptionLegacyDomainStrategyOptions)
 			}
-			server = dialOptions.DomainResolver.Server
+			servers = dialOptions.DomainResolver.Server
 			dnsQueryOptions = adapter.DNSQueryOptions{
-				Transport:              transport,
+				ServerStrategy:         dialOptions.DomainResolver.ServerStrategy,
 				Strategy:               strategy,
 				Timeout:                time.Duration(dialOptions.DomainResolver.Timeout),
 				DisableCache:           dialOptions.DomainResolver.DisableCache,
@@ -103,22 +106,35 @@ func NewWithOptions(options Options) (N.Dialer, error) {
 				RewriteTTL:             dialOptions.DomainResolver.RewriteTTL,
 				ClientSubnet:           dialOptions.DomainResolver.ClientSubnet.Build(netip.Prefix{}),
 			}
+			if len(transports) == 1 {
+				dnsQueryOptions.Transport = transports[0]
+			} else if len(transports) > 1 {
+				dnsQueryOptions.Transports = transports
+			}
 			resolveFallbackDelay = time.Duration(dialOptions.FallbackDelay)
 		} else if options.DirectResolver {
 			return nil, E.New("missing domain resolver for domain server address")
 		} else {
-			if defaultOptions.DomainResolver != "" {
+			if len(defaultOptions.DomainResolvers) > 0 {
 				dnsQueryOptions = defaultOptions.DomainResolveOptions
-				transport, loaded := dnsTransport.Transport(defaultOptions.DomainResolver)
-				if !loaded {
-					return nil, E.New("default domain resolver not found: " + defaultOptions.DomainResolver)
+				transports := make([]adapter.DNSTransport, 0, len(defaultOptions.DomainResolvers))
+				for _, server := range defaultOptions.DomainResolvers {
+					transport, loaded := dnsTransport.Transport(server)
+					if !loaded {
+						return nil, E.New("default domain resolver not found: ", server)
+					}
+					transports = append(transports, transport)
 				}
-				dnsQueryOptions.Transport = transport
+				if len(transports) == 1 {
+					dnsQueryOptions.Transport = transports[0]
+				} else {
+					dnsQueryOptions.Transports = transports
+				}
 				resolveFallbackDelay = time.Duration(dialOptions.FallbackDelay)
 			} else {
 				transports := dnsTransport.Transports()
 				if len(transports) < 2 {
-					dnsQueryOptions.Transport = dnsTransport.Default()
+					dnsQueryOptions.Transport = dnsTransport.Defaults()[0]
 				} else if options.NewDialer {
 					return nil, E.New("missing domain resolver for domain server address")
 				} else {
@@ -137,7 +153,7 @@ func NewWithOptions(options Options) (N.Dialer, error) {
 			options.Context,
 			dialer,
 			dialOptions.Detour == "" && !dialOptions.TCPFastOpen,
-			server,
+			servers,
 			dnsQueryOptions,
 			resolveFallbackDelay,
 		)