+++ daemon/started_service.go
@@ -1279,7 +1279,7 @@ func (s *StartedService) StartNetworkQualityTest(
 		return err
 	}
 
-	resolvedDialer := dialer.NewResolveDialer(boxService.ctx, outbound, true, "", adapter.DNSQueryOptions{}, 0)
+	resolvedDialer := dialer.NewResolveDialer(boxService.ctx, outbound, true, nil, adapter.DNSQueryOptions{}, 0)
 	httpClient := networkquality.NewHTTPClient(resolvedDialer)
 	defer httpClient.CloseIdleConnections()
 
@@ -1325,7 +1325,7 @@ func (s *StartedService) StartSTUNTest(
 		return err
 	}
 
-	resolvedDialer := dialer.NewResolveDialer(boxService.ctx, outbound, true, "", adapter.DNSQueryOptions{}, 0)
+	resolvedDialer := dialer.NewResolveDialer(boxService.ctx, outbound, true, nil, adapter.DNSQueryOptions{}, 0)
 
 	result, stunErr := stun.Run(stun.Options{
 		Server:  request.Server,